requirements.txt
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.8.3
gunicorn==19.4.3
psycopg2==2.6.1
-e git+https://github.com/pybrain/pybrain.git#egg=pybrain
requests==2.7.0
static3==0.6.1
whitenoise==3.0
scipy==0.17.0
numpy==1.11.0
django-crispy-forms==1.6.0
django-picklefield==0.3.2
boto3==1.3.1
